+++ trunk/src/mstl/m_bitset_iterator.h
@@ -1,7 +1,7 @@
 // ======================================================================
 // Author : $Author$
-// Version: $Revision: 1 $
-// Date   : $Date: 2011-05-04 00:04:08 +0000 (Wed, 04 May 2011) $
+// Version: $Revision: 61 $
+// Date   : $Date: 2011-06-30 15:34:21 +0000 (Thu, 30 Jun 2011) $
 // Url    : $URL$
 // ======================================================================
 
@@ -31,7 +31,7 @@ struct bitset_iterator
 	typedef bitfield::reference				reference;
 	typedef bitfield::value_type				value_type;
 	typedef bitset_iterator						iterator;
-	typedef size_t									size_type;
+	typedef unsigned								size_type;
 
 	// structors
 	bitset_iterator();
@@ -74,7 +74,7 @@ struct bitset_const_iterator
 	typedef mstl::bitfield<unsigned long>	bitfield;
 	typedef bool									reference;
 	typedef bitset_const_iterator				const_iterator;
-	typedef size_t									size_type;
+	typedef unsigned								size_type;
 
 	// structors
 	bitset_const_iterator();
